[DEBT-043-D14] Split analysis/scanner.py into focused modules
- Extracted MarketClosedError, ScanFilter, ScanResult to _scanner_models.py (46 LoC)
- Extracted MarketStatusVerifier to _verifier.py (117 LoC)
- Reduced scanner.py to 304 LoC (was 439 LoC)
- Re-exports preserve backwards-compatible public API
- All quality gates pass: pre-commit, pytest (1052 passed)
- Acceptance criteria: 1/1 complete (D14 done)
